Move webgpu_helpers to common.
Consolidate the `webgpu_helpers.{cc|h}` files into the common folder.
